Downgrade bolt_on warning message from printing a message to just a code comment.
As is, this will be printed on every python execution which imports tf_privacy where this module is not available (including transitive dependencies, i.e. via tf_federated). The warning is not actionable for users and so is strictly noise.

Worse, because this uses print() rather than logging, this is often the last line of output when running unit tests. It's easy to mistakenly think that a unit test failure was caused by this warning.

This change removes the log statement entirely, since it is noisy and not actionable.

PiperOrigin-RevId: 274198105