More stable loop termination criterion in fraction RDP order computation.
Previously the loop sometimes terminated too early, resulting in underestimates of the RDP at some orders. Now it will run to convergence in most cases, and in case of too many iterations, it will return an RDP of inf at that order, guaranteeing that the resulting epsilon is a true upper bound.

PiperOrigin-RevId: 668593123